+++ cli/src/main/java/org/eclipse/hono/cli/AbstractCliClient.java
@@ -15,7 +15,10 @@ package org.eclipse.hono.cli;
 
 import java.util.Arrays;
 import java.util.List;
+import java.util.Objects;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.env.Environment;
 
@@ -27,23 +30,45 @@ import io.vertx.core.Vertx;
  */
 public abstract class AbstractCliClient {
 
+    /**
+     * A logger to be shared with subclasses.
+     */
+    protected final Logger LOG = LoggerFactory.getLogger(getClass());
+
+    /**
+     * The vert.x instance to run on.
+     */
     protected Vertx vertx;
+    /**
+     * The vert.x context to run on.
+     */
     protected Context ctx;
+    /**
+     * The Spring Boot profiles that are active.
+     */
     protected List<String> activeProfiles;
 
+    /**
+     * Sets the Spring environment.
+     * 
+     * @param env The environment.
+     * @throws NullPointerException if environment is {@code null}.
+     */
     @Autowired
     public final void setActiveProfiles(final Environment env) {
+        Objects.requireNonNull(env);
         activeProfiles = Arrays.asList(env.getActiveProfiles());
     }
 
     /**
-     * Sets the vertx instance.
+     * Sets the vert.x instance.
      * 
-     * @param vertx The vertx instance.
+     * @param vertx The vert.x instance.
+     * @throws NullPointerException if vert.x is {@code null}.
      */
     @Autowired
     public final void setVertx(final Vertx vertx) {
-        this.vertx = vertx;
+        this.vertx = Objects.requireNonNull(vertx);
         this.ctx = vertx.getOrCreateContext();
     }