+++ cli/src/main/java/org/eclipse/hono/cli/AmqpSend.java
@@ -13,12 +13,18 @@
 package org.eclipse.hono.cli;
 
 import java.io.PrintWriter;
-import java.util.concurrent.CountDownLatch;
+import java.util.Objects;
+import java.util.Optional;
+import java.util.concurrent.CancellationException;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.CompletionException;
 
 import javax.annotation.PostConstruct;
 
 import org.apache.qpid.proton.message.Message;
+import org.eclipse.hono.config.ClientConfigProperties;
 import org.eclipse.hono.util.Strings;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
@@ -27,6 +33,7 @@ import io.vertx.core.Future;
 import io.vertx.proton.ProtonClient;
 import io.vertx.proton.ProtonClientOptions;
 import io.vertx.proton.ProtonConnection;
+import io.vertx.proton.ProtonDelivery;
 import io.vertx.proton.ProtonHelper;
 import io.vertx.proton.ProtonSender;
 import io.vertx.proton.sasl.impl.ProtonSaslPlainImpl;
@@ -38,104 +45,112 @@ import io.vertx.proton.sasl.impl.ProtonSaslPlainImpl;
 @Profile("amqp-adapter-cli")
 public class AmqpSend extends AbstractCliClient {
 
+    private ClientConfigProperties properties = new ClientConfigProperties();
+
+    /**
+     * The address to set on the message being sent.
+     */
+    @Value(value = "${message.address}")
+    private String messageAddress;
+
+    /**
+     * The payload to send.
+     */
+    @Value(value = "${message.payload}")
+    private String payload;
+
     private ProtonConnection adapterConnection;
 
-    @Value(value = "${message.address}")
-    protected String messageAddress;
-
-    @Value(value = "${amqp.host}")
-    protected String amqpHost;
-
-    @Value(value = "${amqp.port}")
-    protected int amqpPort;
-
-    @Value(value = "${username}")
-    protected String username;
-
-    @Value(value = "${password}")
-    protected String password;
-
-    @Value(value = "${payload}")
-    protected String payload;
+    /**
+     * Sets the configuration properties to use for connecting
+     * to the AMQP adapter.
+     * 
+     * @param props The properties.
+     * @throws NullPointerException if properties are {@code null}.
+     */
+    @Autowired(required = false)
+    public void setClientConfig(final ClientConfigProperties props) {
+        this.properties = Objects.requireNonNull(props);
+    }
 
     @PostConstruct
     void start() {
 
-        final Message message = ProtonHelper.message(payload);
-        message.setAddress(messageAddress);
+        final CompletableFuture<ProtonDelivery> messageSent = new CompletableFuture<>();
+        final Message message = ProtonHelper.message(messageAddress, payload);
 
-        final CountDownLatch sent = new CountDownLatch(1);
-        connectToAdapter().setHandler(result -> {
-            final PrintWriter pw = new PrintWriter(System.out);
-            if (result.succeeded()) {
-                final ProtonSender sender = adapterConnection.createSender(null);
-                sender.openHandler(remoteAttach -> {
-                    if (remoteAttach.succeeded()) {
-                        sender.send(message, delivery -> {
-                            // Logs the delivery state to the console
-                            pw.println("\n" + delivery.getRemoteState() + "\n");
-                            pw.flush();
-
-                            sender.close();
-                            if (adapterConnection != null) {
-                                adapterConnection.close();
-                            }
-                            sent.countDown();
-                        });
-                    }
-                }).open();
-
-            } else {
-                pw.println(result.cause());
-                pw.flush();
-            }
+        connectToAdapter()
+        .compose(con -> {
+            adapterConnection = con;
+            final Future<ProtonSender> senderTracker = Future.future();
+            final ProtonSender sender = adapterConnection.createSender(null);
+            sender.openHandler(senderTracker);
+            sender.open();
+            return senderTracker;
+        })
+        .map(sender -> {
+            sender.send(message, delivery -> messageSent.complete(delivery));
+            return sender;
+        })
+        .otherwise(t -> {
+            messageSent.completeExceptionally(t);
+            return null;
         });
 
+        final PrintWriter pw = new PrintWriter(System.out);
+
         try {
-            sent.await();
-            System.exit(0);
-        } catch (InterruptedException e) {
+            final ProtonDelivery delivery = messageSent.join();
+            // Logs the delivery state to the console
+            pw.println("\n" + delivery.getRemoteState() + "\n");
+            pw.flush();
+        } catch (CompletionException e) {
+            pw.println(e.getCause());
+            pw.flush();
+        } catch (CancellationException e) {
             // do-nothing
         }
+        if (adapterConnection != null) {
+            adapterConnection.close();
+        }
+        System.exit(0);
     }
 
     // ----------------------------------< Vertx-proton >---
 
-    private Future<Void> connectToAdapter() {
-        final Future<Void> result = Future.future();
+    private Future<ProtonConnection> connectToAdapter() {
+
+        final Future<ProtonConnection> connectAttempt = Future.future();
         final ProtonClientOptions options = new ProtonClientOptions();
         final ProtonClient client = ProtonClient.create(vertx);
-        if (!Strings.isNullOrEmpty(username) && !Strings.isNullOrEmpty(password)) {
+
+        Optional.ofNullable(properties.getAmqpHostname()).ifPresent(s -> options.setVirtualHost(s));
+
+        if (!Strings.isNullOrEmpty(properties.getUsername()) && !Strings.isNullOrEmpty(properties.getPassword())) {
             // SASL PLAIN authc.
             options.addEnabledSaslMechanism(ProtonSaslPlainImpl.MECH_NAME);
-            client.connect(options, amqpHost, amqpPort, username, password, conAttempt -> {
-                if (conAttempt.failed()) {
-                    result.fail(conAttempt.cause());
-                } else {
-                    adapterConnection = conAttempt.result();
-                    adapterConnection.openHandler(remoteOpen -> {
-                        if (remoteOpen.succeeded()) {
-                            result.complete();
-                        }
-                    }).open();
-                }
-            });
+
+            LOG.info("connecting to AMQP adapter using SASL PLAIN [host: {}, port: {}, username: {}]",
+                    properties.getHost(), properties.getPort(), properties.getUsername());
+
+            client.connect(
+                    options,
+                    properties.getHost(),
+                    properties.getPort(),
+                    properties.getUsername(),
+                    properties.getPassword(),
+                    connectAttempt);
         } else {
             // SASL ANONYMOUS authc.
-            client.connect(amqpHost, amqpPort, conAttempt -> {
-                if (conAttempt.failed()) {
-                    result.fail(conAttempt.cause());
-                } else {
-                    adapterConnection = conAttempt.result();
-                    adapterConnection.openHandler(remoteOpen -> {
-                        if (remoteOpen.succeeded()) {
-                            result.complete();
-                        }
-                    }).open();
-                }
-            });
+            client.connect(properties.getHost(), properties.getPort(), connectAttempt);
         }
 
-        return result;
+        return connectAttempt
+                .compose(unopenedConnection -> {
+                    final Future<ProtonConnection> con = Future.future();
+                    unopenedConnection.openHandler(con);
+                    unopenedConnection.open();
+                    return con;
+                });
     }
 }